+++ oci/objectstorage_replication_policy_test.go
@@ -66,6 +66,7 @@ func TestObjectStorageReplicationPolicyResource_basic(t *testing.T) {
 	singularDatasourceName := "data.oci_objectstorage_replication_policy.test_replication_policy"
 
 	var resId string
+
 	resource.Test(t, resource.TestCase{
 		PreCheck: func() { testAccPreCheck(t) },
 		Providers: map[string]terraform.ResourceProvider{
@@ -83,6 +84,16 @@ func TestObjectStorageReplicationPolicyResource_basic(t *testing.T) {
 					resource.TestCheckResourceAttrSet(resourceName, "destination_region_name"),
 					resource.TestCheckResourceAttr(resourceName, "name", "mypolicy"),
 					resource.TestCheckResourceAttrSet(resourceName, "namespace"),
+
+					func(s *terraform.State) (err error) {
+						resId, err = fromInstanceState(s, resourceName, "id")
+						if isEnableExportCompartment, _ := strconv.ParseBool(getEnvSettingWithDefault("enable_export_compartment", "true")); isEnableExportCompartment {
+							if errExport := testExportCompartmentWithResourceName(&resId, &compartmentId, resourceName); errExport != nil {
+								return errExport
+							}
+						}
+						return err
+					},
 				),
 			},
 
@@ -103,16 +114,6 @@ func TestObjectStorageReplicationPolicyResource_basic(t *testing.T) {
 					resource.TestCheckResourceAttrSet(datasourceName, "replication_policies.0.status"),
 					resource.TestCheckResourceAttrSet(datasourceName, "replication_policies.0.status_message"),
 					resource.TestCheckResourceAttrSet(datasourceName, "replication_policies.0.time_created"),
-
-					func(s *terraform.State) (err error) {
-						resId, err = fromInstanceState(s, resourceName, "id")
-						if isEnableExportCompartment, _ := strconv.ParseBool(getEnvSettingWithDefault("enable_export_compartment", "false")); isEnableExportCompartment {
-							if errExport := testExportCompartmentWithResourceName(&resId, &compartmentId, resourceName); errExport != nil {
-								return errExport
-							}
-						}
-						return err
-					},
 				),
 			},