+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 4.8.0 (Unreleased)
+
+### Fixed
+- Fixed typo in load balancer shape. Issue [#1252](https://github.com/terraform-providers/terraform-provider-oci/issues/1252)
+
 ## 4.7.0 (December 16, 2020)
 
 ### Added
@@ -10,6 +15,10 @@
 
 ### Deprecated
 - Data source `oci_database_autonomous_database_wallet` is being deprecated in favor of resource `oci_database_autonomous_database_wallet` 
+
+### Fixed
+- Fix `idcs_access_token` as `required` now. Note that this will break current import scenario, user need to add `idcs_access_token` to Terraform config.
+[Github issue 1264](https://github.com/terraform-providers/terraform-provider-oci/issues/1264)
 
 ## 4.6.0 (December 09, 2020)